+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 [Back](./README.md)
 
+## v2.1.0 (Dec 15 2025)
+
+- **fix**: Air pressure device MA10238 had wrong humidity key (error) [#40](https://github.com/CestLaGalere/mobilealerts/issues/40)
+- **fix**: Fix unit in rain flip counter from mm to none. Count history. Needed a separate sensor. [#41](https://github.com/CestLaGalere/mobilealerts/issues/41)
+- **feat**: Add sensor MA10250PRO in device list with existing sensors types [#40](https://github.com/CestLaGalere/mobilealerts/issues/40)
+- **feat**: Add sensor MA10410 weather station [#45](https://github.com/CestLaGalere/mobilealerts/issues/45)
+- **feat**: Add service `Dump Raw API Response` which can be called by user. The service dumps the raw json for all configured devices.
+- **chore**: Adapt readme.md with better log info and describe how to count the rain sensor (with use of Utility Meters)
+
+
 ## v2.0.0 (Dec 3 2025)
 
 - **feat**: Add a device via UI on page Settings -> Device & services -> Integrations -> mobile Alerts, Current YAML configuration still works in parallel, information for migration in readme